JS 133 видео Promise. CRUD на примере to-do Feature: Подтверждение удаления отдельной задачи

+++ practice/propmise/03/src/components/inits/initDeleteTask.js
@@ -1,9 +1,13 @@
 import {deleteTask} from "../../api/index.js";
 import {loadData} from "../index.js";
-import {showError} from "../../utils/notification.js";
+import {showConfirmation, showError} from "../../utils/notification.js";
 
 export function initDeleteTask(deleteButtonEl, task) {
     deleteButtonEl.addEventListener('click', async () => {
+        const isConfirmed = await showConfirmation('Вы уверены, что хотите удалить эту задачу, без возможности восстановления?');
+        if (!isConfirmed) {
+            return
+        }
         try {
             await deleteTask(task.id)
             await loadData();
